requirements.txt
alembic==1.9.4
anyio==3.6.2
asyncpg==0.26.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
databases==0.6.1
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.6
ormar==0.11.3
ormar-postgres-full-text==0.1.0
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.25.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
